Add files via upload
The very first thing we do in our main function is set a dirs variable to a tuple, which are the directions (N,E,S,W). I then created our six rooms which are each dictionaries. The directions are then set for each of the rooms. A list of rooms is created because we want to choose a random room for the tree to be hidden. At the start of the game, you begin with a room and prints a message. And then we begin the loop which is a game loop and the rest of this program is within this loop and once you break out of it, you are done. For example, if the presents are delivered and the room name is 'Chimney', then you are done and prints a congratulatory message. However, if the presents have not been delivered you are given a message, one that is intended to help guide you into the right direction. We then have a while loop created for when the user is either hitting invalid characters or entering a direction in which there is no door. If a valid entry is entered then a message is printed out. This either prints out a message telling you 'cannot go in that direction' or prints out the room name in which you have entered. This game loop continues until the presents are successfully delivered and you are then given the chance to restart or end the game.